chore(arena): move streaming files
